requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
